1. 소개페이지 추가 2. 메인페이지 비워뒀던 차트 추가 3. 프로필 페이지 추가
