+++ package.json
@@ -18,7 +18,9 @@
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-dropzone": "^14.2.3",
     "react-router-dom": "^6.25.0",
+    "recharts": "^2.12.7",
     "tailwindcss": "^3.4.6"
   },
   "devDependencies": {
