+++ src/api/axiosInstance.jsx
@@ -1,0 +1,28 @@
+import axios from "axios";
+
+const API_URL = 'http://localhost:8080';
+
+const axiosInstance = axios.create({
+    baseURL: API_URL,
+    headers: {
+        'Content-Type': 'application/json',
+    },
+});
+
+axiosInstance.interceptors.request.use(
+    (config) => {
+        const token = localStorage.getItem('accessToken');
+        console.log(token);
+        if (token) {
+            config.headers['Authorization'] = `Bearer ${token}`;
+        } else {
+            config.headers['Authorization'] = null;
+        }
+        return config;
+    },
+    (error) => {
+        return Promise.reject(error);
+    }
+);
+
+export default axiosInstance;
