--- src/components/Products.jsx
@@ -1,62 +1,0 @@
-import React from 'react';
-
-const products = [
-    {
-        id: 1,
-        name: '전 세계 생산량',
-        href: '#',
-        price: 'date',
-        imageSrc: '/fresh-grape-leaves.jpg',
-        imageAlt: '전 세계 PE, PP, PS 폐기량',
-    },
-    {
-        id: 2,
-        name: '전 세계 폐기량',
-        href: '#',
-        price: 'date',
-        imageSrc: '/stuffed-grape-leaves.jpg',
-        imageAlt: '전 세계 PP, PE PS ',
-    },
-    {
-        id: 3,
-        name: '재활용 가격',
-        href: '#',
-        price: 'date',
-        imageSrc: '/stuffed-grape-leaves.jpg',
-        imageAlt: 'PP, PE PS 가격 변동',
-    },
-];
-
-export function Products() {
-    return (
-        <div className="bg-white">
-            <div className="max-w-2xl mx-auto py-16 px-4 sm:py-24 sm:px-6 lg:max-w-7xl lg:px-8">
-                <h2 className="text-2xl font-extrabold tracking-tight text-gray-900">PE, PP, PS 추이</h2>
-                <div className="mt-6 grid grid-cols-1 gap-y-10 gap-x-6 sm:grid-cols-2 lg:grid-cols-4 xl:gap-x-8">
-                    {products.map((product) => (
-                        <div key={product.id} className="group relative">
-                            <div className="w-full min-h-80 bg-gray-200 aspect-w-1 aspect-h-1 rounded-md overflow-hidden group-hover:opacity-75 lg:h-80 lg:aspect-none">
-                                <img
-                                    src={product.imageSrc}
-                                    alt={product.imageAlt}
-                                    className="w-full h-full object-center object-cover lg:w-full lg:h-full"
-                                />
-                            </div>
-                            <div className="mt-4 flex justify-between">
-                                <div>
-                                    <h3 className="text-sm text-gray-700">
-                                        <a href={product.href}>
-                                            <span aria-hidden="true" className="absolute inset-0" />
-                                            {product.name}
-                                        </a>
-                                    </h3>
-                                </div>
-                                <p className="text-sm font-medium text-gray-900">{product.price}</p>
-                            </div>
-                        </div>
-                    ))}
-                </div>
-            </div>
-        </div>
-    );
-}
