--- src/components/Profile.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Profile = () => {
-    return (
-        <div className="max-w-md mx-auto mt-10 p-6 bg-white rounded-lg shadow-xl">
-            <h2 className="text-2xl font-bold mb-4">User Profile</h2>
-            <p>This is the user profile page. You can add more details here.</p>
-        </div>
-    );
-};
-
-export default Profile;
