+++ src/components/home/Main.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export function Main() {
+    return (
+        <div className="relative w-full h-screen bg-cover bg-center"
+             style={{backgroundImage: "url('/src/assets/main.png')"}}>
+        </div>
+    );
+}
